requirements.txt
pathlib
ruamel-yaml
roboflow
pymongo
gridfs